+++ plugin/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## unreleased
 
+- EXL-284 - Added assets to overwrite user menu templates
+- EXL-284 - Added install step to remove rights for all unnecessary menu items, for every role, except super-admin
+- EXL-284 - Added hook that redefine menus, to remove menu items that were not removed based on role rights
 - EXL-285 - Added htaccess based redirect to https 
 - EXL-274 - Added import tab to the config page
 - EXL-274 - config page to the plugin